Move some files to src/server/
