+++ src/server.h
@@ -34,7 +34,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "util/basic_macros.h"
 #include "util/metricsbackend.h"
 #include "serverenvironment.h"
-#include "clientiface.h"
+#include "server/clientiface.h"
 #include "chatmessage.h"
 #include "sound.h"
 #include "translation.h"
