+++ src/server/rollback.cpp
